perfect

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: vborysov <vborysov@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/06 13:32:51 by vborysov          #+#    #+#             */
-/*   Updated: 2025/11/06 13:33:07 by vborysov         ###   ########.fr       */
+/*   Updated: 2025/11/10 19:48:36 by vborysov         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
-	int	result;
+	size_t	result;
 
 	result = 0;
 	while (str[result])
